Adds TwitsService
